problem 1090 try 16(answer)

+++ learning.c
@@ -2,13 +2,13 @@
 
 int main(){
 
-    int a,b,c,i=0;
+    long long int a,b,c,i=0;
     scanf("%d %d %d", &a, &b, &c);
     for(i=1; i < c; i++){
-        a=a*b;
+        a = a*b;
     }
 
-    printf("%d", a);
+    printf("%lld", a);
 
     return 0;
 }
